# [ET-VK] Simplify conv2d weight prepacking (>2x pipeline-creation speedup)